kola: Enable SELinux as early as possible
We never really tested SELinux because we enabled it after boot while
normally it would be permanently enabled even during (re)boot.
We need to enable it via Ignition. Since this won't work with old
releases due to policy problems, introduce a flag that the old scripts
branches can pass.
Note: If tests differ between early and non-early enabling I would
rather disable SELinux for those cases and add a comment if and under
what future conditions it can be reenabled. The alternative would be to
only make them run with the new early mode but this means we reduce test
coverage for Stable which is not a good idea.