kola: Extend test to cover recreation of files in /etc
For files like /etc/resolv.conf we want it to be recreated because the
tmpfiles rule would do this. For a systemd bug workaround we filtered
out some rules but even if we add them back, we still want to ensure
that not only the file will always be recreated but that it also is not
frozen as upcopy because we prefer the overlay lowerdir as source for
updates.
Extend the /etc overlayfs test to ensure recreation of the file and to
check against upcopies.