Skip zfs test in GHA and for release tests
GHA doesn't upload artifacts to bincache so the test won't be able to fetch the
sysext. Skip it by checking for version.txt for the tested version on bincache.
This requires modifying the test skipping logic to fetch VERSION from
os-release (instead of VERSION_ID) but semver can handle this because it
encodes the string after the '+' in a "metadata" and ignores it when making
comparisons between versions.

Signed-off-by: Jeremi Piotrowski <jpiotrowski@microsoft.com>