kola/tests/tang: Ignore interfaces with no IPs
We use Equinix Metal machines in our CI that have bonded network interfaces,
which means we end up with nics with no ips. The tang setup code needs to
handle this. We will also want to rework the code to provide tang in the same
way as we do for etcd.

Signed-off-by: Jeremi Piotrowski <jpiotrowski@microsoft.com>