kola/tests/tang: Refactor test to use socket in qemu namespace
The current version of the test creates a host socket which is affected by host
iptables settings. This is why both tang tests are currently failing on our
hosted infra. Rework the test to create a listener in the network namespace of
the kola flight, which avoids host networking issues. We still have to
dynamically find a port to avoid test case port collision, because kola runs
all (qemu) tests inside the same network namespace.

Signed-off-by: Jeremi Piotrowski <jpiotrowski@microsoft.com>