kola: Increase amount of disk corruption in verity test
For a particular image build the filesystem seemed to be more resistant
to the corruption test, be it due to the corruption hitting other
structures or some non-flushable cache covering more of the corrupted
area. One can still trigger the verity panic by overwriting 100 MB
instead of 10 MB.
Increase the amount of zeros written by the corruption test.