kola: Run qemu and swtpm with the machine folder as working dir
The long UNIX socket path caused problems and the workaround was to
shorten the test names but a real solution is to ensure that the path
is always short. This can be done by setting the current working
directory for the swtpm process to the machine folder which also holds
the socket and, in addition, setting the current working directory for
the qemu process to the same folder. Then we can use 'tpm/socket' as
path which is ensured to be below the limit of 108 characters.