platform: Update azure to new SDK
This commit replaces old resource clients with their counterparts from
the new SDK. Most of the changes made were rather straightforward. The
larger changes were made in storage account handling and blob
uploading. We stopped generating shared keys for storage accounts, so
for uploading we expect the authenticated user to have an RBAC
permission that allows creating containers and upload blobs already
assigned. This can be done by, for example, assigning "Storage Blob
Data Owner" role at a subscription level. Uploading the blobs is now
delegated to the azure-vhd-utils package.

Another thing that has changed is authentication to Azure - the whole
thing is replaced with a default authentication method provided by the
azidentity package. So support for azure profile and azure credentials
JSON files is dropped. More about the default authentication method
can be read on azidentity documentation website:

https://pkg.go.dev/github.com/Azure/azure-sdk-for-go/sdk/azidentity#section-readme

Some functionality that was unused and was difficult to port (image
sharing, replication and unreplication, some OS image stuff) was
dropped.