.github: Add an action for building docker images on-demand
There isn't a clear way to tell when the image should be rebuilt
automatically. In theory it should be when the relevant debian package
gets an update or when a new Debian version is released. In practice,
these are annoying to track, so the action is to be invoked on demand.