api/aws: remove details during the import task status
The details (progress and status message) are not provided by AWS at the beginning of the import. Here's the
details:
```json
{
  DiskImageSize: 0,
  Format: "RAW",
  SnapshotId: "",
  Status: "active",
  UserBucket: {
    S3Bucket: "flatcar-kola-ami-import-us-east-1",
    S3Key: "tormath1/amd64-usr/tormath1-flatcar-stable/flatcar_production_ami_image.bin"
  }
}
```

Let's just drop it as we don't use it, this does not impact the
execution of the program. If required, we can still access to the task
via its ID using the AWS CLI.

Signed-off-by: Mathieu Tortuyaux <mtortuyaux@microsoft.com>