kola: Add checks for AVC messages
Kola does some checks after every test run. They are just regexps,
that will cause the test to fail if they match. The regexps are being
run over the contents of the console output and journal. We add
another check to match the AVC denial line ("avc: denied { … } …").
There is a flag to skip the check in case a test wants to actually
generate an AVC and check for it itself.