kola: Add an option for specifying image version
When we are running our tests on Github or in Jenkins, we always know
what version of Flatcar the image used for testing has. So instead of
figuring it out by creating a throw-away cluster, make it possible to
provide the image version through a command-line flag. This commit
drops the fast-track code-path that was skipping the version query, so
in order to avoid the overhead of creating a cluster, passing the
version through the command line is required.

This also fixes an issue of SELinux AVC not being ignored on older
versions of the image if getting the image version from the cluster
was for some reason skipped.