kola/nvidia: Skip tests at runtime
We used to do it this way, then I moved the skipping to SkipFunc. Unfortunately
this doesn't work well for vendor-testing/azure.sh in flatcar/scripts which
passes 'cl.internet cl.misc.nvidia' to one GPU VM size and to one Gen1 VM size.
It would be much more complex to straighten this out in bash, because the
arguments are passed into other helper functions so restore runtime skipping.

Switch to creating the machine from the testcase (ClusterSize=0), so that the
skip happens before the machine is created, saving time.

Signed-off-by: Jeremi Piotrowski <jpiotrowski@microsoft.com>