kola/tests: Fix tests following the kernel move into /usr
cl.update.reboot still behaves as before.

cl.verity is mostly the same, but there is a shorter GRUB retry loop
rather than a longer reboot loop because GRUB is unable to load the
kernel from the corrupted /usr.

cl.update.badverity causes GRUB's load script GPG verification to fail
rather than the kernel's root hash verification, but the end result
should be the same. The security is hard to break.

coreos.update.badusr is now skipped because signed load scripts are now
used, we lack the GPG key to sign a new one, and the kernel is supposed
to be loaded from /usr.

Signed-off-by: James Le Cuirot <jlecuirot@microsoft.com>