update: boot: add cl.update.payload-boot-part-too-small test case
This test case makes sure that the Flatcar upgrade fails if the
update payload contains an initrd that has a bigger size than the
free space in the /boot partition.

Currently, the /boot partition free space is around 70 MB.

The test case is Writing 20 MB on the /boot partition before starting
the upgrade, so that it makes sure there is not enough space for the
upgrade to complete. The stable channel has now around one vmlinuz plus
around 15MB free (that s why 20MB has been chosen).

NOTE: this test's expected result is to make sure that the upgrade
FAILS if there is not enough /boot partition space, and at the
same time, verifies that Flatcar can still boot with its
currrent version after the failure.

Signed-off-by: Adrian Vladu <avladu@cloudbasesolutions.com>