kubeadm: exclude Akamai tests
We reached the monthly credit cap offered by Akamai by retrying too many
times those failing tests.

Kubernetes tests are failing because the OS boots in a 4Gb image which
is not enough to pull all kubernetes components.

There is an on-going discussion with Akamai at the moment, but we need
to exclude Kubernetes tests for Akamai to avoid spending all our credits
here.

Signed-off-by: Mathieu Tortuyaux <mtortuyaux@microsoft.com>