kola: Avoid dynamic offset calculation in badusr test
When /usr was ext2, we could query the filesystem size to get the verity
offset. However, since we have a fixed /usr size anyway and with btrfs
we use a hardcoded offset, switch the test to use the same offset so
that we can avoid shipping e2size in the initrd.

Signed-off-by: Kai Lueke <kailuke@microsoft.com>