kola: Change cl.update.payload-boot-part-too-small to always fill /boot
The test case used a fixed 20 MB file to fill the /boot disk so that the
second kernel copy wouldn't fit. However, with the reduced minimal
initrd the kernel fits four times and thus the single 20 MB file
doesn't prevent the update.
Change the dd call to write as much as possible so that the disk is
always filled afterwards.

Signed-off-by: Kai Lueke <kailuke@microsoft.com>