platform: Increase image size for GCE and AWS
The GCE test VMs had 12 GB storage hardcoded and AWS 8 but after
increasing the /boot, /usr, and /oem partitions, the new image size is
not 8 GB as with the raw image but 13 GB because the rootfs is set to 6
GB in the VM disk layout, and for AWS we also switched to the same
layout as other VMs while before it had a very small rootfs.
Increase the hardcoded VM storage size to 16 GB to compensate for the
almost 4 GB partition increase.

Signed-off-by: Kai Lueke <kailuke@microsoft.com>