kola: Extend sysext update test
The sysext update test covered OEM extensions, first with the fallback
download and then with the locally passed extension - however, this was
still able to use the fallback download and thus we missed a regression
in ue-rs.
Set up a "unknown" Flatcar extension which can't exist on bincache to
make sure we cover locally passed Flatcar extensions.

Signed-off-by: Kai Lueke <kailuke@microsoft.com>