kola: Add test for kernel to have built-in microcode firmware files
Recently we didn't have CPU microcode updates anymore.
Add a test that checks that the kernel has the firmware built-in. This
doesn't test if they actually get applied because that's not possible in
QEMU but anyway that should be part of the kernel upstream QA. Here we
rather want to prevent package use flags or similar from impacting the
kernel build system's inclusion of the files.

Signed-off-by: Kai Lueke <kailuke@microsoft.com>