kola: Support default confext for /etc
When we switch from a custom overlay mount for /etc to using
systemd-confext there are a few minor differences with unmounting /etc
from a new namespace for inspection and when checking extension paths
that are now resolved when being symlinks.
Also add a new test for making sure that the confext/sysext skipping
works.

Signed-off-by: Kai Lueke <kailuke@microsoft.com>